+++ GroceryStores/GroceryStores/Pages/Index.cshtml.cs
@@ -20,67 +20,50 @@ namespace GroceryStores.Pages
         }
         public void OnGet()
         {
+            //Business Owners data from the API
+            string BusinessOwnerJsonString = GetData("https://data.cityofchicago.org/resource/r5kz-chrr.json");
+            BusinessOwner[] allBusinessOwners = BusinessOwner.FromJson(BusinessOwnerJsonString);
+            ViewData["allBusinessOwners"] = allBusinessOwners;
+
+            //Grocery stores data from te API
+            string GroceryStoreJsonString = GetData("https://data.cityofchicago.org/resource/53t8-wyrc.json");
+            GroceryStore[] allGroceryStores = GroceryStore.FromJson(GroceryStoreJsonString);
+            ViewData["allGroceryStores"] = allGroceryStores;
+
+            //Map to store the the zipcode as key and grocery store as Value
+            IDictionary<long, GroceryStore> groceryStoresMap = new Dictionary<long, GroceryStore>();
+
+            //List to store filtered business owners
+            List<BusinessOwner> businessOwnerList = new List<BusinessOwner>();
+
+
+            //Iterating through Grocery stores and adding them to the groceryStoresMap
+            foreach (GroceryStore grocery in allGroceryStores)
+            {
+                if (!groceryStoresMap.ContainsKey(grocery.ZipCode))
+                {
+                    groceryStoresMap.Add(grocery.ZipCode, grocery);
+                }
+            }
+
+            //Filtering the Business Owners and adding thefiltered records to businessOwnerList
+            foreach (BusinessOwner businessRec in allBusinessOwners)
+            {
+                if (groceryStoresMap.ContainsKey(businessRec.ZipCode))
+                {
+                    businessOwnerList.Add(businessRec);
+                }
+            }
+            ViewData["businessOwners"] = businessOwnerList;
+        }
+
+        //This generic method is used to get the JsonString by calling an API
+        public string GetData(string url)
+        {
             using (WebClient webClient = new WebClient())
             {
-                string businessOwnerEndPoint = "https://data.cityofchicago.org/resource/r5kz-chrr.json";
-                string BusinessOwnerJsonString = webClient.DownloadString(businessOwnerEndPoint);
-                BusinessOwner[] allBusinessOwners = BusinessOwner.FromJson(BusinessOwnerJsonString);
-                ViewData["allBusinessOwners"] = allBusinessOwners;
-
-                string groceryStoreEndPoint = "https://data.cityofchicago.org/resource/53t8-wyrc.json";
-                string GroceryStoreJsonString = webClient.DownloadString(groceryStoreEndPoint);
-                GroceryStore[] allGroceryStores = GroceryStore.FromJson(GroceryStoreJsonString);
-                ViewData["allGroceryStores"] = allGroceryStores;
-
-                IDictionary<long, GroceryStore> groceryStoresMap = new Dictionary<long, GroceryStore>();
-                List<BusinessOwner> businessOwnerList = new List<BusinessOwner>();
-
-                foreach (GroceryStore grocery in allGroceryStores)
-                {
-                    if (!groceryStoresMap.ContainsKey(grocery.ZipCode))
-                    {
-                        groceryStoresMap.Add(grocery.ZipCode, grocery);
-                    }                    
-                }
-
-
-                foreach (BusinessOwner businessRec in allBusinessOwners)
-                {
-                    if (groceryStoresMap.ContainsKey(businessRec.ZipCode))
-                    {
-                        businessOwnerList.Add(businessRec);
-                    }
-                }
-
-                /*IDictionary<long, BusinessOwner> businessOwnersMap = new Dictionary<long, BusinessOwner>();
-
-                List<GroceryStore> groceryStoresOwners = new List<GroceryStore>();
-
-                foreach(BusinessOwner businessOwner in allBusinessOwners){
-                    businessOwnersMap.Add(businessOwner.LicenseId, businessOwner);
-                }
-
-
-                foreach (GroceryStore groceryStore in allGroceryStores)
-                {
-                    if (businessOwnersMap.ContainsKey(groceryStore.LicenseId))
-                    {
-                        groceryStoresOwners.Add(groceryStore);
-                    }
-                }*/
-
-
-                //foreach (BusinessOwner businessOwner in allBusinessOwners)
-                //{
-                //    foreach (var groceryStore in allGroceryStores)
-                //    {
-                //        if (groceryStore.LicenseId == businessOwner.LicenseId)
-                //        {
-                //            groceryStoresOwners.AddRange(groceryStore.LicenseId, groceryStore);
-                //        }
-                //    }
-                //}
-                ViewData["businessOwners"] = businessOwnerList;
+                string jsonString = webClient.DownloadString(url);
+                return jsonString;
             }
         }
     }
